+++ app-backend/src/__tests__/handler.test.ts
@@ -5,12 +5,13 @@
 * Redistribution and use in source and binary forms, with or
 * without modification, are not permitted.
 */
+
 import {graphqlHandler, playgroundHandler} from '../handler';
 
-it('graphqlHandler should be a function', () => {
-  expect(typeof graphqlHandler).toBe('function');
+it('graphqlHandler should be a function', (): void => {
+    expect(typeof graphqlHandler).toBe('function');
 });
 
-it('playgroundHandler should be a function', () => {
-  expect(typeof playgroundHandler).toBe('function');
+it('playgroundHandler should be a function', (): void => {
+    expect(typeof playgroundHandler).toBe('function');
 });