+++ app-client/src/__mocks__/localStorage.js
@@ -16,4 +16,4 @@ class LocalStorageMock {
 	}
 }
 
-global.localStorage = new LocalStorageMock;
+global.localStorage = new LocalStorageMock();