+++ app-backend/src/handler.ts
@@ -8,9 +8,9 @@
 
 import lambdaPlayground from 'graphql-playground-middleware-lambda';
 import {Handler, Context, Callback, APIGatewayEvent} from 'aws-lambda';
-import {graphqlLambda, graphiqlLambda, LambdaHandler } from 'apollo-server-lambda';
-import { ITypeDefinitions } from 'graphql-tools/dist/Interfaces';
-import {makeExecutableSchema } from 'graphql-tools';
+import {graphqlLambda, graphiqlLambda, LambdaHandler} from 'apollo-server-lambda';
+import {ITypeDefinitions} from 'graphql-tools/dist/Interfaces';
+import {makeExecutableSchema} from 'graphql-tools';
 import {resolvers} from './resolvers';
 
 const typeDefs: ITypeDefinitions = require('./query.graphql');
@@ -22,6 +22,7 @@ const myGraphQLSchema = makeExecutableSchema({
 });
 
 export const graphqlHandler: Handler = (event: APIGatewayEvent, context: Context, callback: Callback) => {
+    context.callbackWaitsForEmptyEventLoop = false;
     const handler: LambdaHandler = graphqlLambda({ schema: myGraphQLSchema, tracing: true });
     return handler(event, context, callback);
 };