+++ CHANGELOG.md
@@ -22,5 +22,6 @@
 
 ### Added
 
-- Added AuthRoute Component
-- Added Homepage
+- AuthRoute Component
+- Homepage
+- The UserDataMapper models, schemas and the resolvers.