+++ CHANGELOG.md
@@ -25,3 +25,4 @@
 - AuthRoute Component
 - Homepage
 - The UserDataMapper models, schemas and the resolvers.
+- `eslint-config-react-app` to the dev dependencies.